% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_split.R
\name{codon_split}
\alias{codon_split}
\title{Codon split}
\usage{
codon_split(sequence, start = 1)
}
\arguments{
\item{sequence}{That's the sequence, your input, that you want to split in codons.}

\item{start}{That's the position within the sequence where the splitting should start.}
}
\value{
The output is a codon list in order and by fragments of 3.
}
\description{
Codon split
}
