% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_concatenation.R
\name{codons_concanation}
\alias{codons_concanation}
\title{Condons Concanation}
\usage{
codons_concanation(codons)
}
\arguments{
\item{codons}{here is where the codons are stored}
}
\value{
a table of concanated codons is returned
}
\description{
Condons Concanation
}
