% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_concatenation.R
\name{codons_concatenation}
\alias{codons_concatenation}
\title{Codons Concatenation}
\usage{
codons_concatenation(codons)
}
\arguments{
\item{codons}{here is where the codons are stored}
}
\value{
a table of concatenated codons is returned
}
\description{
Codons Concatenation
}
