% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_builder.R
\name{dna_builder}
\alias{dna_builder}
\title{DNA Builder :)}
\usage{
dna_builder(seq_length)
}
\arguments{
\item{seq_length}{input the desired sequence length}
}
\value{
the output is a random RNA sequence of desired length
}
\description{
DNA Builder :)
}
