% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{DNA To RNA}
\usage{
DNA_to_RNA(DNA_string)
}
\arguments{
\item{DNA_string}{'A string in DNA code'}
}
\value{
'A string in RNA code'
}
\description{
This function converts a DNA string to RNA
}
