% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nucleotides.R
\name{generate_nucleotides}
\alias{generate_nucleotides}
\title{Generate a random nucleotide sequence given a size input}
\usage{
generate_nucleotides(seq_size)
}
\arguments{
\item{seq_size}{the length of the sequence to generate}
}
\value{
nucleotide sequence as string
}
\description{
Generate a random nucleotide sequence given a size input
}
