% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_aminoacid_count.R
\name{plot_by_aminoacid_count}
\alias{plot_by_aminoacid_count}
\title{Plot By Aminoacid Count}
\usage{
plot_by_aminoacid_count(aminoacid)
}
\arguments{
\item{protein}{"A string of aminoacids"}
}
\value{
"bar plot by aminoacid count"
}
\description{
Plot By Aminoacid Count
}
