% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Changing DNA into RNA}
\usage{
dna_to_rna(codons)
}
\arguments{
\item{codons}{A character string or character vector containing DNA sequence(s)
compoed of the letters A, T, C and G}
}
\value{
A character string or character vector with all "T" bases replaced by "U".
}
\description{
This function takes one or more DNA sequences and replaces all thymine (T) bases with uracil (U), converting the sequence(s) into RNA form.
}
\examples{
dna_to_rna("ATTGCT")
dna_to_rna(c("TTTAAA", "GGGCCC"))
}
